+++ log.c
@@ -173,6 +173,71 @@ log_dbg_mode(int mode)
 
 
 /*
+ * Master key log.  Logs master keys in SSLKEYLOGFILE format.
+ * Uses a logger thread.
+ */
+
+logger_t *masterkey_log = NULL;
+static int masterkey_fd = -1;
+static char *masterkey_fn = NULL;
+
+static int
+log_masterkey_preinit(const char *logfile)
+{
+	masterkey_fd = open(logfile, O_WRONLY|O_APPEND|O_CREAT, DFLT_FILEMODE);
+	if (masterkey_fd == -1) {
+		log_err_printf("Failed to open '%s' for writing: %s (%i)\n",
+		               logfile, strerror(errno), errno);
+		return -1;
+	}
+	if (!(masterkey_fn = realpath(logfile, NULL))) {
+		log_err_printf("Failed to realpath '%s': %s (%i)\n",
+		              logfile, strerror(errno), errno);
+		close(masterkey_fd);
+		masterkey_fd = -1;
+		return -1;
+	}
+	return 0;
+}
+
+static int
+log_masterkey_reopencb(void)
+{
+	close(masterkey_fd);
+	masterkey_fd = open(masterkey_fn, O_WRONLY|O_APPEND|O_CREAT,
+	                    DFLT_FILEMODE);
+	if (masterkey_fd == -1) {
+		log_err_printf("Failed to open '%s' for writing: %s\n",
+		               masterkey_fn, strerror(errno));
+		free(masterkey_fn);
+		masterkey_fn = NULL;
+		return -1;
+	}
+	return 0;
+}
+
+/*
+ * Do the actual write to the open master key log file descriptor.
+ */
+static ssize_t
+log_masterkey_writecb(UNUSED void *fh, const void *buf, size_t sz)
+{
+	if (write(masterkey_fd, buf, sz) == -1) {
+		log_err_printf("Warning: Failed to write to masterkey log:"
+		               " %s\n", strerror(errno));
+		return -1;
+	}
+	return sz;
+}
+
+static void
+log_masterkey_fini(void)
+{
+	close(masterkey_fd);
+}
+
+
+/*
  * Connection log.  Logs a one-liner to a file-based connection log.
  * Uses a logger thread.
  */
@@ -983,6 +1048,17 @@ log_preinit(opts_t *opts)
 			goto out;
 		}
 	}
+	if (opts->masterkeylog) {
+		if (log_masterkey_preinit(opts->masterkeylog) == -1)
+			goto out;
+		if (!(masterkey_log = logger_new(log_masterkey_reopencb,
+		                                 NULL, NULL,
+		                                 log_masterkey_writecb, NULL,
+		                                 log_exceptcb))) {
+			log_masterkey_fini();
+			goto out;
+		}
+	}
 	if (opts->certgendir) {
 		if (!(cert_log = logger_new(NULL, NULL, NULL, log_cert_writecb,
 		                            NULL, log_exceptcb)))
@@ -1005,12 +1081,17 @@ out:
 	if (cert_log) {
 		logger_free(cert_log);
 	}
+	if (masterkey_log) {
+		log_masterkey_fini();
+		logger_free(masterkey_log);
+	}
 	return -1;
 }
 
 /*
  * Close all file descriptors opened by log_preinit; used in privsep parent.
- * Only undo content and connect log, leave error and debug log functional.
+ * Only undo content, connect and masterkey logs, leave error and debug log
+ * functional.
  */
 void
 log_preinit_undo(void)
@@ -1022,6 +1103,10 @@ log_preinit_undo(void)
 	if (connect_log) {
 		log_connect_fini();
 		logger_free(connect_log);
+	}
+	if (masterkey_log) {
+		log_masterkey_fini();
+		logger_free(masterkey_log);
 	}
 }
 
@@ -1039,6 +1124,9 @@ log_init(opts_t *opts, proxy_ctx_t *ctx, int clisock1, int clisock2)
 	if (!opts->debug) {
 		err_shortcut_logger = 1;
 	}
+	if (masterkey_log)
+		if (logger_start(masterkey_log) == -1)
+			return -1;
 	if (connect_log)
 		if (logger_start(connect_log) == -1)
 			return -1;
@@ -1072,6 +1160,8 @@ log_fini(void)
 
 	if (cert_log)
 		logger_leave(cert_log);
+	if (masterkey_log)
+		logger_leave(masterkey_log);
 	if (content_log)
 		logger_leave(content_log);
 	if (connect_log)
@@ -1081,6 +1171,8 @@ log_fini(void)
 
 	if (cert_log)
 		logger_join(cert_log);
+	if (masterkey_log)
+		logger_join(masterkey_log);
 	if (content_log)
 		logger_join(content_log);
 	if (connect_log)
@@ -1090,6 +1182,8 @@ log_fini(void)
 
 	if (cert_log)
 		logger_free(cert_log);
+	if (masterkey_log)
+		logger_free(masterkey_log);
 	if (content_log)
 		logger_free(content_log);
 	if (connect_log)
@@ -1097,6 +1191,8 @@ log_fini(void)
 	if (err_log)
 		logger_free(err_log);
 
+	if (masterkey_log)
+		log_masterkey_fini();
 	if (content_log)
 		log_content_file_fini();
 	if (connect_log)
@@ -1113,6 +1209,9 @@ log_reopen(void)
 {
 	int rv = 0;
 
+	if (masterkey_log)
+		if (logger_reopen(masterkey_log) == -1)
+			rv = -1;
 	if (content_log)
 		if (logger_reopen(content_log) == -1)
 			rv = -1;