+++ log.h
@@ -45,6 +45,18 @@ void log_dbg_mode(int);
 #define LOG_DBG_MODE_NONE 0
 #define LOG_DBG_MODE_ERRLOG 1
 
+extern logger_t *masterkey_log;
+#define log_masterkey_printf(fmt, ...) \
+        logger_printf(masterkey_log, NULL, 0, (fmt), __VA_ARGS__)
+#define log_masterkey_print(s) \
+        logger_print(masterkey_log, NULL, 0, (s))
+#define log_masterkey_write(buf, sz) \
+        logger_write(masterkey_log, NULL, 0, (buf), (sz))
+#define log_masterkey_print_free(s) \
+        logger_print_freebuf(masterkey_log, NULL, 0, (s))
+#define log_masterkey_write_free(buf, sz) \
+        logger_write_freebuf(masterkey_log, NULL, 0, (buf), (sz))
+
 extern logger_t *connect_log;
 #define log_connect_printf(fmt, ...) \
         logger_printf(connect_log, NULL, 0, (fmt), __VA_ARGS__)